# openapi.yaml
openapi: 3.1.0
info:
  title: WinMyDispute API
  version: 3.1.0
  description: |
    GPT-facing API for the WinMyDispute free-preview and premium-upgrade flow.
    The free tier returns a likely reason-code match, a confidence read, and exactly one strategy tip.
    Premium output unlocks after a one-time Stripe Checkout payment tied to the user's email and can produce
    a professional dispute packet as JSON plus downloadable PDF, Word-compatible RTF, or text report output.
  contact:
    name: Daniel Neville
    email: pureplays@icloud.com
    url: https://pure-plays.com
  license:
    name: Proprietary
    url: https://pure-plays.com/terms
servers:
  - url: https://api.pure-plays.com
    description: Production custom domain
  - url: https://winmydispute-api.onrender.com
    description: Render deployment
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/disputes/preview:
    post:
      tags: [disputes]
      summary: Return the free dispute preview
      description: |
        This is the GPT's free-value endpoint. It returns the likely dispute reason,
        a confidence score, exactly one preview tip, and the list of premium features.
        Inputs may contain flexible date and amount formats plus lightly misspelled consumer descriptions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DisputeInput"
      responses:
        "200":
          description: Free preview response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/generate-letter:
    post:
      tags: [disputes]
      summary: Generate the full premium dispute kit
      description: |
        Requires a paid email license plus either a signed premium access token or the matching
        Stripe checkout session ID. When unpaid, the endpoint returns HTTP 402 with a checkout URL.
        When paid, the endpoint returns the full premium package in one response, including issuer submission guidance,
        an exhibit packet, and a submission-ready action plan.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DisputeInput"
                - type: object
                  required: [email]
      responses:
        "200":
          description: Full premium response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PremiumResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before premium generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/generate-report-document:
    post:
      tags: [disputes]
      summary: Generate a premium dispute report document
      description: |
        Requires a paid email license plus either a signed premium access token or the matching
        Stripe checkout session ID. Produces a downloadable, professionally formatted report package.
        Supported formats are PDF, DOCX, Word-compatible RTF, and plain text. The endpoint can also
        generate a second redacted copy for privacy-safe sharing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DisputeInput"
                - type: object
                  required: [email]
      responses:
        "200":
          description: Report document metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDocumentResponse"
        "202":
          description: Report generation queued asynchronously
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAcceptedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before premium generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/generate-submission-bundle:
    post:
      tags: [disputes]
      summary: Generate a downloadable submission bundle
      description: |
        Requires a paid email license plus either a signed premium access token or the matching
        Stripe checkout session ID. Produces a ZIP package with the generated report,
        submission plan, exhibit index, and a compact case summary.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DisputeInput"
                - type: object
                  required: [email]
      responses:
        "200":
          description: Submission bundle metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionBundleResponse"
        "202":
          description: Bundle generation queued asynchronously
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAcceptedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before bundle generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/jobs/{jobId}:
    get:
      tags: [disputes]
      summary: Retrieve an asynchronous job result
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
        - name: premiumAccessToken
          in: query
          required: false
          schema:
            type: string
          description: Signed premium access token returned by /auth/check-license after payment.
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
          description: Matching Stripe Checkout session ID for token exchange.
        - name: checkoutSessionId
          in: query
          required: false
          schema:
            type: string
          description: Alias for sessionId.
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "402":
          description: Upgrade required before job access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpgradeRequiredResponse"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    BadRequest:
      description: Invalid request input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Request limit exceeded for the current rate-limit window
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitError"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: A valid email address is required.
        requestId:
          type: string
          format: uuid
          nullable: true
      required: [error]
...
    EvidenceItem:
      oneOf:
        - type: string
        - type: object
          additionalProperties: true
          properties:
            summary:
              type: string
            description:
              type: string
            text:
              type: string
            filename:
              type: string
            fileType:
              type: string
            type:
              type: string
            extractedText:
              type: string
    DisputeInput:
      type: object
      additionalProperties: true
      properties:
        email:
          type: string
          format: email
          description: Optional for preview, required for premium generation and license lookup.
        premiumAccessToken:
          type: string
          description: Signed premium access token returned by /auth/check-license after payment.
        checkoutSessionId:
          type: string
          description: Matching Stripe Checkout session ID that can be exchanged for a premium access token.
        sessionId:
          type: string
          description: Alias for checkoutSessionId.
        network:
          type: string
          enum: [visa, mastercard, amex, discover]
        issuer:
          type: string
          example: Chase
        cardholderName:
          type: string
          example: Jane Doe
        addressLine1:
          type: string
          example: 123 Main St
        addressLine2:
          type: string
          example: Miami, FL 33101
        phone:
          type: string
          example: "3055551212"
        merchantName:
          type: string
          example: Example Merchant
        merchantVertical:
          type: string
          example: subscription
        transactionAmount:
          oneOf:
            - type: number
            - type: string
          example: $1,234.56
          description: Flexible amount input. The API normalizes common U.S. and international display formats.
        transactionDate:
          type: string
          example: 03/01/2026
          description: Flexible date input. The API normalizes common slash, dash, dot, and month-name formats.
        description:
          type: string
          description: Primary dispute summary used for reason-code matching.
          example: I cancelled before renewal but the merchant chargd me anyway.
        evidenceSummary:
          type: string
          example: I have cancellation emails and screenshots of the account page.
        evidenceItems:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        timelineItems:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        merchantRebuttalConcerns:
          type: array
          items:
            $ref: "#/components/schemas/EvidenceItem"
        desiredOutcome:
          type: string
          example: Reverse the charge and confirm the billing correction in writing.
        tone:
          type: string
          enum: [formal, assertive, concise, empathetic]
          example: formal
        lengthPreference:
          type: string
          enum: [short, standard, detailed]
          example: detailed
        outputFormat:
          type: string
          enum: [pdf, docx, word-compatible-rtf, text]
          example: pdf
        async:
          type: boolean
          description: When true on supported premium endpoints, the server returns a queued job instead of waiting for completion.
          example: true
        redactionMode:
          type: string
          enum: [none, standard, strict]
          description: Optional privacy mode for generated report artifacts.
          example: standard
        includeRedactedVersion:
          type: boolean
          description: When true, the report-document endpoint returns both a full copy and a redacted copy.
          example: true
      required: [description]
...
    PreviewResponse:
      type: object
      properties:
        reasonCode:
          type: string
          nullable: true
          example: 13.2
        network:
          type: string
          example: visa
        merchantVertical:
          type: string
          nullable: true
          example: subscription
        confidence:
          type: integer
          minimum: 0
          maximum: 100
          example: 82
        confidenceNote:
          type: string
          example: This confidence measures how strongly the facts match the likely dispute reason. It is not a guaranteed outcome score.
        previewTip:
          type: string
          description: Exactly one free preview tip.
          example: Lead with the exact date and proof that you cancelled before the billing date.
        successEstimate:
          $ref: "#/components/schemas/SuccessEstimate"
        reviewFlags:
          type: array
          items:
            $ref: "#/components/schemas/ReviewFlag"
        upgradeRequired:
          type: boolean
          example: true
        premiumFeatures:
          type: array
          items:
            type: string
          example:
            - full-letter
            - full-strategy-set
            - evidence-checklist
            - cfpb-summary
      required: [reasonCode, network, confidence, confidenceNote, previewTip, successEstimate, reviewFlags, upgradeRequired, premiumFeatures]
...
    ReviewFlag:
      type: object
      properties:
        code:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        message:
          type: string
        recommendation:
          type: string
      required: [code, severity, message, recommendation]
    SuccessEstimate:
      type: object
      properties:
        estimatedSuccessRate:
          type: number
          minimum: 0
          maximum: 1
        modelType:
          type: string
          enum: [heuristic, blended-historical-and-heuristic]
        sampleSize:
          type: integer
        rationale:
          type: string
      required: [estimatedSuccessRate, modelType, sampleSize, rationale]
...
    CaseVersion:
      type: object
      properties:
        versionId:
          type: string
        createdAt:
          type: string
          format: date-time
        stage:
          type: string
      required: [versionId, createdAt, stage]
...
    JobStatusResponse:
      type: object
      properties:
        jobId:
          type: string
        kind:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        email:
          type: string
          format: email
          nullable: true
        caseId:
          type: string
          nullable: true
        requestId:
          type: string
          nullable: true
        premiumAccessToken:
          type: string
          nullable: true
        input:
          type: object
          additionalProperties: true
        result:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        errorMessage:
          oneOf:
            - type: string
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        completedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
      required: [jobId, kind, status, input, createdAt, updatedAt]
    JobAcceptedResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, running]
        email:
          type: string
          format: email
        premiumAccessToken:
          type: string
          nullable: true
        requestId:
          type: string
          nullable: true
        storedFiles:
          type: array
          items:
            $ref: "#/components/schemas/StoredFileSummary"
      required: [jobId, status, email]
...
    PremiumResponse:
      type: object
      properties:
        email:
          type: string
          format: email
        premiumAccessToken:
          type: string
          nullable: true
        reasonCode:
          type: string
          nullable: true
        network:
          type: string
        merchantVertical:
          type: string
          nullable: true
        confidence:
          type: integer
          minimum: 0
          maximum: 100
        letter:
          type: string
        strategySet:
          $ref: "#/components/schemas/StrategySet"
        evidenceChecklist:
          type: array
          items:
            type: string
        evidencePacket:
          $ref: "#/components/schemas/EvidencePacket"
        exhibitPacket:
          $ref: "#/components/schemas/ExhibitPacket"
        issuerGuidance:
          $ref: "#/components/schemas/IssuerGuidance"
        submissionPlan:
          $ref: "#/components/schemas/SubmissionPlan"
        cfpbSummary:
          type: string
        successEstimate:
          $ref: "#/components/schemas/SuccessEstimate"
        reviewFlags:
          type: array
          items:
            $ref: "#/components/schemas/ReviewFlag"
        filingReadiness:
          $ref: "#/components/schemas/FilingReadiness"
        documentPreferences:
          $ref: "#/components/schemas/DocumentPreferences"
        caseFile:
          $ref: "#/components/schemas/CaseFileSummary"
        caseVersion:
          $ref: "#/components/schemas/CaseVersion"
      required:
        - email
        - reasonCode
        - network
        - confidence
        - letter
        - strategySet
        - evidenceChecklist
        - evidencePacket
        - exhibitPacket
        - issuerGuidance
        - submissionPlan
        - cfpbSummary
        - successEstimate
        - reviewFlags
        - filingReadiness
        - documentPreferences
        - caseFile
        - caseVersion
    ReportDocumentResponse:
      type: object
      properties:
        premiumAccessToken:
          type: string
          nullable: true
        kind:
          type: string
          enum: [artifact]
        fileId:
          type: string
        format:
          type: string
          example: docx
        filename:
          type: string
          example: winmydispute-report-1711200000000.docx
        url:
          type: string
          format: uri-reference
          example: /api/v1/artifacts/abc123?expires=1711203600000&signature=deadbeef
        expiresAt:
          type: string
        contentType:
          type: string
          example: application/vnd.openxmlformats-officedocument.wordprocessingml.document
        variant:
          type: string
          enum: [full, redacted]
        redactionMode:
          type: string
          enum: [none, standard, strict]
        requestedRedactionMode:
          type: string
          enum: [none, standard, strict]
        includeRedactedVersion:
          type: boolean
        artifacts:
          type: array
          items:
            $ref: "#/components/schemas/ReportArtifact"
        redactedArtifact:
          oneOf:
            - $ref: "#/components/schemas/ReportArtifact"
            - type: "null"
        caseFile:
          $ref: "#/components/schemas/CaseFileSummary"
        caseVersion:
          $ref: "#/components/schemas/CaseVersion"
      required: [kind, fileId, format, filename, url, expiresAt, contentType, variant, redactionMode, requestedRedactionMode, includeRedactedVersion, artifacts, caseFile, caseVersion]
    SubmissionBundleResponse:
      type: object
      properties:
        premiumAccessToken:
          type: string
          nullable: true
        kind:
          type: string
          enum: [artifact]
        fileId:
          type: string
        format:
          type: string
          enum: [zip]
        filename:
          type: string
        url:
          type: string
          format: uri-reference
        expiresAt:
          type: string
        contentType:
          type: string
        bundleItems:
          type: array
          items:
            type: string
        sourceArtifacts:
          type: array
          items:
            $ref: "#/components/schemas/ReportArtifact"
        requestedOutputFormat:
          type: string
        caseFile:
          $ref: "#/components/schemas/CaseFileSummary"
        caseVersion:
          $ref: "#/components/schemas/CaseVersion"
      required: [kind, fileId, format, filename, url, expiresAt, contentType, bundleItems, sourceArtifacts, requestedOutputFormat, caseFile, caseVersion]
...
    UpgradeRequiredResponse:
      type: object
      properties:
        upgradeRequired:
          type: boolean
          example: true
        checkoutUrl:
          type: string
          format: uri
          example: https://checkout.stripe.com/pay/cs_test_123
        checkoutSessionId:
          type: string
          example: cs_test_123
        message:
          type: string
          example: Premium access requires a one-time $6.99 payment.
      required: [upgradeRequired, checkoutUrl, checkoutSessionId, message]